Proof of Concept of init-images
This command:

  fn init --image=fnproj/java-init:latest

Will `docker run` that image, piping the output to `tar -x`. The expected
result is that a template of the function will be created, along with a
func.yaml which is missing `Name` and `Version`.  These are then added by
the cli, so that the function is ready to `fn run`.

Future work is to remove the hard-coded templates in the cli, and make

  fn init --runtime=java

equivalent to the line above. Once this is done for all *supported* runtimes,
the others (php, lambda-node, rust, etc) can be removed and LanguageHelper
is then unnecessary.

Co-Author: vittoria.placidi@oracle.com